+++ src/graphql/event/kerberos.rs
@@ -155,7 +155,7 @@ impl BlocklistKerberos {
 
     /// Client Name
     async fn client_name(&self) -> &[String] {
-        &self.inner.client_name
+        &self.inner.cname
     }
 
     /// Realm
@@ -170,7 +170,7 @@ impl BlocklistKerberos {
 
     /// Service Name
     async fn service_name(&self) -> &[String] {
-        &self.inner.service_name
+        &self.inner.sname
     }
 
     /// MITRE Tactic