+++ src/graphql/node/input.rs
@@ -54,7 +54,7 @@ pub struct AgentInput {
 impl From<AgentInput> for review_database::Agent {
     fn from(input: AgentInput) -> Self {
         Self {
-            node: u32::MAX,
+            node_id: u32::MAX,
             key: input.key,
             kind: input.kind.into(),
             status: input.status.into(),
@@ -90,7 +90,7 @@ pub struct ExternalServiceInput {
 impl From<ExternalServiceInput> for review_database::ExternalService {
     fn from(input: ExternalServiceInput) -> Self {
         Self {
-            node: u32::MAX,
+            node_id: u32::MAX,
             key: input.key,
             kind: input.kind.into(),
             status: input.status.into(),
@@ -188,7 +188,7 @@ pub(super) fn create_draft_update(
                     };
 
                     Ok(review_database::Agent {
-                        node: u32::MAX,
+                        node_id: u32::MAX,
                         key: new_agent.key,
                         kind: new_agent.kind.into(),
                         status: new_agent.status.into(),
@@ -218,7 +218,7 @@ pub(super) fn create_draft_update(
                     };
 
                     Ok(review_database::ExternalService {
-                        node: u32::MAX,
+                        node_id: u32::MAX,
                         key: new_external_service.key,
                         kind: new_external_service.kind.into(),
                         status: new_external_service.status.into(),